+++ python/GafferUI/PathListingWidget.py
@@ -600,13 +600,23 @@ class PathListingWidget( GafferUI.Widget ) :
 	def __buttonPress( self, widget, event ) :
 
 		self.__updateSelectionInButtonRelease = False
-		if event.buttons != event.Buttons.Left :
-			return False
 
 		# Get model index under cursor.
 
 		index = self.__indexAt( event.line.p0 )
 		if index is None :
+			return False
+
+		# Delegate the click to the PathColumn, if it wants it.
+
+		if self.getColumns()[index.column()].buttonPressSignal()(
+			self.__pathForIndex( index ), self, event
+		) :
+			return True
+
+		# Otherwise, perform expansion and selection as appropriate.
+
+		if event.buttons != event.Buttons.Left :
 			return False
 
 		# Do expansion/collapsing if the arrow was clicked on. QTreeView doesn't
@@ -651,23 +661,30 @@ class PathListingWidget( GafferUI.Widget ) :
 
 	def __buttonRelease( self, widget, event ) :
 
-		if not self.__updateSelectionInButtonRelease :
+		index = self.__indexAt( event.line.p0 )
+		if index is None :
 			return False
+
+		if self.__updateSelectionInButtonRelease :
+			self.__singleSelect( index )
+			return True
+		else :
+			return self.getColumns()[index.column()].buttonReleaseSignal()(
+				self.__pathForIndex( index ), self, event
+			)
+
+	def __buttonDoubleClick( self, widget, event ) :
 
 		index = self.__indexAt( event.line.p0 )
 		if index is None :
 			return False
 
-		self.__singleSelect( index )
-		return True
+		if self.getColumns()[index.column()].buttonDoubleClickSignal()(
+			self.__pathForIndex( index ), self, event
+		) :
+			return True
 
-	def __buttonDoubleClick( self, widget, event ) :
-
-		if event.buttons != event.Buttons.Left :
-			return False
-
-		index = self.__indexAt( event.line.p0 )
-		if index is not None :
+		if event.buttons == event.Buttons.Left :
 			self.__activated( index )
 			return True