+++ src/GafferUI/PathColumn.cpp
@@ -78,6 +78,21 @@ PathColumn::PathColumnSignal &PathColumn::changedSignal()
 	return m_changedSignal;
 }
 
+PathColumn::ButtonSignal &PathColumn::buttonPressSignal()
+{
+	return m_buttonPressSignal;
+}
+
+PathColumn::ButtonSignal &PathColumn::buttonReleaseSignal()
+{
+	return m_buttonReleaseSignal;
+}
+
+PathColumn::ButtonSignal &PathColumn::buttonDoubleClickSignal()
+{
+	return m_buttonDoubleClickSignal;
+}
+
 //////////////////////////////////////////////////////////////////////////
 // StandardPathColumn
 //////////////////////////////////////////////////////////////////////////